requirements.txt
pymupdf==1.24.0
pytesseract==0.3.10
Pillow==10.3.0
opencv-python==4.9.0.80
invoice2data==0.4.4
pandas==2.2.2
click==8.1.7
joblib==1.4.0